Move Custom Windows views
