+++ ArchipelagoParser/ViewModel/RandoStatsViewModel.cs
@@ -1,16 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Dynamic;
+﻿using System.Collections.ObjectModel;
 using System.IO;
 using System.Net.Http;
-using System.Text;
 using System.Text.Json;
 using System.Windows;
-using NoNiDev.ArchipelagoParser.Views;
+using NoNiDev.ArchipelagoParser.Views.CustomWindows;
 using NoNiDev.CallAPI.RandoStat;
 using NoNiDev.SpoilerArchipelagoParser;
-using NoNiDev.SpoilerArchipelagoParser.SOHOptions;
 
 namespace NoNiDev.ArchipelagoParser.ViewModel
 {
