+++ ArchipelagoParser/Views/CustomWindows/PopUpWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="NoNiDev.ArchipelagoParser.Views.PopUpWindow"
+﻿<Window x:Class="NoNiDev.ArchipelagoParser.Views.CustomWindows.PopUpWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
