+++ ArchipelagoParser/Views/CustomWindows/PopUpWindow.xaml.cs
@@ -1,16 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
+﻿using System.Windows;
 
-namespace NoNiDev.ArchipelagoParser.Views
+namespace NoNiDev.ArchipelagoParser.Views.CustomWindows
 {
     /// <summary>
     /// Interaction logic for PopUpWindow.xaml
